Add static well-known file to verify Logseq on Flathub
This will add the verified badget to the Flathub app page
https://flathub.org/apps/com.logseq.Logseq

Verification is a process by which Flathub can confirm that an app is
published by its original developer or a third party approved by the developer.
This is done by verifying the developer's identity using a link to their
website or to their account on a source code hosting site.

See https://docs.flathub.org/docs/for-app-authors/verification